Parasites Web Application/cs36-main-master/requirements.txt
certifi==2020.12.5
Django==2.2.17
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
wincertstore==0.2
django-ckeditor==6.0.0